+++ graph/adapter-output-spring-data-neo4j-ogm/src/main/kotlin/eu/tib/orkg/prototype/statements/adapter/output/neo4j/spring/SpringDataNeo4jPredicateAdapter.kt
@@ -68,7 +68,7 @@ class SpringDataNeo4jPredicateAdapter(
         ]
     )
     override fun save(predicate: Predicate) {
-        neo4jRepository.save(predicate.toNeo4jPredicate())
+        neo4jRepository.save(predicate.toNeo4jPredicate(neo4jRepository))
     }
 
     override fun nextIdentity(): ThingId {