+++ graph/adapter-output-spring-data-neo4j-ogm/src/main/kotlin/eu/tib/orkg/prototype/statements/adapter/output/neo4j/spring/SpringDataNeo4jStatementAdapter.kt
@@ -13,19 +13,15 @@ import eu.tib.orkg.prototype.statements.adapter.output.neo4j.spring.internal.Neo
 import eu.tib.orkg.prototype.statements.adapter.output.neo4j.spring.internal.Neo4jStatement
 import eu.tib.orkg.prototype.statements.adapter.output.neo4j.spring.internal.Neo4jStatementIdGenerator
 import eu.tib.orkg.prototype.statements.adapter.output.neo4j.spring.internal.Neo4jStatementRepository
-import eu.tib.orkg.prototype.statements.adapter.output.neo4j.spring.internal.Neo4jThing
 import eu.tib.orkg.prototype.statements.api.BundleConfiguration
 import eu.tib.orkg.prototype.statements.api.RetrieveStatementUseCase
-import eu.tib.orkg.prototype.statements.domain.model.Class
 import eu.tib.orkg.prototype.statements.domain.model.GeneralStatement
 import eu.tib.orkg.prototype.statements.domain.model.Literal
 import eu.tib.orkg.prototype.statements.domain.model.Predicate
 import eu.tib.orkg.prototype.statements.domain.model.Resource
 import eu.tib.orkg.prototype.statements.domain.model.StatementId
-import eu.tib.orkg.prototype.statements.domain.model.Thing
 import eu.tib.orkg.prototype.statements.domain.model.ThingId
 import eu.tib.orkg.prototype.statements.spi.OwnershipInfo
-import eu.tib.orkg.prototype.statements.spi.PredicateRepository
 import eu.tib.orkg.prototype.statements.spi.ResourceContributor
 import eu.tib.orkg.prototype.statements.spi.StatementRepository
 import java.util.*
@@ -37,13 +33,12 @@ typealias PredicateLookupTable = Map<ThingId, Predicate>
 
 @Component
 class SpringDataNeo4jStatementAdapter(
-    private val neo4jRepository: Neo4jStatementRepository,
-    private val neo4jStatementIdGenerator: Neo4jStatementIdGenerator,
-    private val neo4jResourceRepository: Neo4jResourceRepository,
-    private val neo4jPredicateRepository: Neo4jPredicateRepository,
-    private val neo4jLiteralRepository: Neo4jLiteralRepository,
-    private val neo4jClassRepository: Neo4jClassRepository,
-    private val predicateRepository: PredicateRepository,
+    val neo4jRepository: Neo4jStatementRepository,
+    val neo4jStatementIdGenerator: Neo4jStatementIdGenerator,
+    val neo4jResourceRepository: Neo4jResourceRepository,
+    val neo4jPredicateRepository: Neo4jPredicateRepository,
+    val neo4jLiteralRepository: Neo4jLiteralRepository,
+    val neo4jClassRepository: Neo4jClassRepository,
 ) : StatementRepository {
     override fun nextIdentity(): StatementId {
         // IDs could exist already by manual creation. We need to find the next available one.
@@ -199,45 +194,15 @@ class SpringDataNeo4jStatementAdapter(
     ): Optional<GeneralStatement> =
         neo4jRepository.findBySubjectIdAndPredicateIdAndObjectId(subjectId, predicateId, objectId).map { it.toStatement() }
 
-    private fun Neo4jStatement.toStatement(): GeneralStatement = GeneralStatement(
-        id = statementId!!,
-        subject = subject!!.toThing(),
-        predicate = predicateRepository.findByPredicateId(ThingId(predicateId!!.value)).get(),
-        `object` = `object`!!.toThing(),
-        createdAt = createdAt!!,
-        createdBy = createdBy
+    private fun Neo4jStatement.toStatement() = toStatement(neo4jPredicateRepository)
+
+    private fun GeneralStatement.toNeo4jStatement() = toNeo4jStatement(
+        neo4jStatementRepository = neo4jRepository,
+        neo4jClassRepository = neo4jClassRepository,
+        neo4jLiteralRepository = neo4jLiteralRepository,
+        neo4jPredicateRepository = neo4jPredicateRepository,
+        neo4jResourceRepository = neo4jResourceRepository
     )
-
-    private fun Neo4jStatement.toStatement(lookupTable: PredicateLookupTable): GeneralStatement = GeneralStatement(
-        id = statementId!!,
-        subject = subject!!.toThing(),
-        predicate = lookupTable[ThingId(predicateId!!.value)]
-            ?: throw IllegalStateException("Predicate $predicateId not found in lookup table. This is a bug."),
-        `object` = `object`!!.toThing(),
-        createdAt = createdAt!!,
-        createdBy = createdBy
-    )
-
-    private fun GeneralStatement.toNeo4jStatement(): Neo4jStatement =
-        // Need to fetch the internal ID of a (possibly) existing entity to prevent creating a new one.
-        neo4jRepository.findByStatementId(id!!).orElse(Neo4jStatement()).apply {
-            statementId = this@toNeo4jStatement.id
-            subject = this@toNeo4jStatement.subject.toNeo4jThing()
-            `object` = this@toNeo4jStatement.`object`.toNeo4jThing()
-            predicateId = this@toNeo4jStatement.predicate.id.toPredicateId()
-            createdBy = this@toNeo4jStatement.createdBy
-            createdAt = this@toNeo4jStatement.createdAt
-        }
-
-    private fun Thing.toNeo4jThing(): Neo4jThing =
-        // The purpose of this method is of technical nature, as the OGM requires a reference to the start and end node.
-        // With direct access to the database, this becomes obsolete, because queries can use the ID directly.
-        when (this) {
-            is Class -> neo4jClassRepository.findByClassId(this.id.toClassId()).get()
-            is Literal -> neo4jLiteralRepository.findByLiteralId(this.id.toLiteralId()).get()
-            is Predicate -> neo4jPredicateRepository.findByPredicateId(this.id.toPredicateId()).get()
-            is Resource -> neo4jResourceRepository.findByResourceId(this.id.toResourceId()).get()
-        }
 
     private fun BundleConfiguration.toApocConfiguration(): Map<String, Any> {
         val conf = mutableMapOf<String, Any>(