+++ graph/adapter-output-spring-data-neo4j-sdn6/src/main/kotlin/eu/tib/orkg/prototype/statements/adapter/output/neo4j/spring/internal/Neo4jClassRepository.kt
@@ -49,4 +49,7 @@ RETURN COUNT(node)""")
     fun findAllByLabelContaining(label: String, pageable: Pageable): Page<Neo4jClass>
 
     fun findByUri(uri: String): Optional<Neo4jClass>
+
+    @Query("""MATCH (c:Class) DETACH DELETE c""")
+    override fun deleteAll()
 }