+++ library/src/main/kotlin/eu/tib/orkg/prototype/statements/application/Exceptions.kt
@@ -128,6 +128,18 @@ class DOIServiceUnavailable : LoggedMessageException {
         super(HttpStatus.SERVICE_UNAVAILABLE, """DOI service returned "$responseMessage" with error response: $errorResponse""")
 }
 
+class InvalidSubclassRelation(childId: ThingId, parentId: ThingId) :
+    SimpleMessageException(HttpStatus.BAD_REQUEST, """The class "$childId" cannot be a subclass of "$parentId"."""")
+
+class ParentClassAlreadyExists(childId: ThingId, parentId: ThingId) :
+    SimpleMessageException(HttpStatus.BAD_REQUEST, """The class "$childId" already has a parent class ($parentId)."""")
+
+class EmptyChildIds :
+    SimpleMessageException(HttpStatus.BAD_REQUEST, "The provided list is empty.")
+
+class ParentClassAlreadyHasChildren(id: ThingId) :
+    SimpleMessageException(HttpStatus.BAD_REQUEST, """The class "$id" already has a child classes."""")
+
 class MissingParameter private constructor(
     override val message: String
 ) : SimpleMessageException(HttpStatus.BAD_REQUEST, message, null) {