+++ rest-api-server/src/integrationTest/kotlin/eu/tib/orkg/prototype/statements/application/ClassControllerIntegrationTest.kt
@@ -382,23 +382,26 @@ class ClassControllerIntegrationTest : RestDocumentationBaseTest() {
             fieldWithPath("unlisted").description("Unlisted Value").optional().ignored()
         )
 
-    private fun classListResponseFields() =
-        listOf(
-            fieldWithPath("id").description("The class ID").optional(),
-            fieldWithPath("label").description("The class label"),
-            fieldWithPath("uri").description("An optional URI to describe the class (RDF)").optional(),
-            fieldWithPath("created_at").description("The class creation datetime"),
-            fieldWithPath("created_by").description("The ID of the user that created the class. All zeros if unknown."),
-            fieldWithPath("description").description("The description of the class, if exists.").optional(),
-            fieldWithPath("_class").optional().ignored(),
-            fieldWithPath("featured").description("Featured Value").optional().ignored(),
-            fieldWithPath("unlisted").description("Unlisted Value").optional().ignored()
-        )
+    companion object RestDoc {
+        fun classListResponseFields() =
+            listOf(
+                fieldWithPath("id").description("The class ID").optional(),
+                fieldWithPath("label").description("The class label"),
+                fieldWithPath("uri").description("An optional URI to describe the class (RDF)").optional(),
+                fieldWithPath("created_at").description("The class creation datetime"),
+                fieldWithPath("created_by").description("The ID of the user that created the class. All zeros if unknown."),
+                fieldWithPath("description").description("The description of the class, if exists.").optional(),
+                fieldWithPath("_class").optional().ignored(),
+                fieldWithPath("featured").description("Featured Value").optional().ignored(),
+                fieldWithPath("unlisted").description("Unlisted Value").optional().ignored()
+            )
+    }
 
     private fun resourceListDetailedResponseFields() =
         responseFields(pageableDetailedFieldParameters())
-            .andWithPrefix("content[].", resourceListInnerResponseFields()
-                    ).andWithPrefix("")
+            .andWithPrefix(
+                "content[].", resourceListInnerResponseFields()
+            ).andWithPrefix("")
 
     fun resourceListInnerResponseFields() = listOf(
         fieldWithPath("id").description("The resource ID"),